handle pending request ui for receiver
